add chat and time systems. split connection test into 2 parts, one that runs the old raw message logger and another that uses the client class. Fix 64 bit long packing and unpacking.

+++ BZFlag.Data/Utils/DynamicOutputBuffer.cs
@@ -139,7 +139,7 @@ namespace BZFlag.Data.Utils
 			BytesUsed += 8;
 		}
 
-		public void WriteInt64(Int32 value)
+		public void WriteInt64(Int64 value)
 		{
 			CheckBuffer(8);
